Move operation mode functions to iwl_op_mode
Bring our fork back in line with upstream iwlwifi in using iwl_op_mode
for operation calls. This allows MVM and DVM to use the same prototype
calls with on-the-fly conversion of the structs. Obviously some explicit
casts had to be added to accomodate for C++ not automatically casting
void to whatever one is trying to use it as.